+++ packages/f-vue-icons/src/components/BaseProviderIcon/BaseProviderIcon.vue
@@ -53,7 +53,7 @@ export default {
     },
     computed: {
         iconType () {
-            return `${this.icon.key}-icon`;
+            return `${this.icon.key}-provider-icon`;
         }
     }
 };