+++ packages/f-vue-icons/src/components/BaseProviderIcon/index.js
@@ -1,0 +1,7 @@
+import BaseProviderIcon from './BaseProviderIcon.vue';
+
+const install = Vue => {
+    Vue.component(BaseProviderIcon.name, BaseProviderIcon);
+};
+
+export default install;